Remove catalog and refactor screenshot testing helper
