+++ .idea/gradle.xml
@@ -22,7 +22,6 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
-            <option value="$PROJECT_DIR$/app-geto-catalog" />
             <option value="$PROJECT_DIR$/benchmarks" />
             <option value="$PROJECT_DIR$/build-logic" />
             <option value="$PROJECT_DIR$/build-logic/convention" />
