--- app-geto-catalog/build.gradle.kts
@@ -1,52 +1,0 @@
-/*
- *
- *   Copyright 2023 Einstein Blanco
- *
- *   Licensed under the GNU General Public License v3.0 (the "License");
- *   you may not use this file except in compliance with the License.
- *   You may obtain a copy of the License at
- *
- *       https://www.gnu.org/licenses/gpl-3.0
- *
- *   Unless required by applicable law or agreed to in writing, software
- *   distributed under the License is distributed on an "AS IS" BASIS,
- *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *   See the License for the specific language governing permissions and
- *   limitations under the License.
- *
- */
-
-plugins {
-    alias(libs.plugins.com.android.geto.application)
-    alias(libs.plugins.com.android.geto.applicationCompose)
-}
-
-android {
-    namespace = "com.android.geto.getocatalog"
-
-    defaultConfig {
-        applicationId = "com.android.geto.getocatalog"
-        versionCode = 1
-        versionName = "0.0.1" // X.Y.Z; X = Major, Y = minor, Z = Patch level
-    }
-
-    buildTypes {
-        release {
-            // To publish on the Play store a private signing key is required, but to allow anyone
-            // who clones the code to sign and run the release variant, use the debug signing key.
-            // TODO: Abstract the signing configuration to a separate file to avoid hardcoding this.
-            signingConfig = signingConfigs.named("debug").get()
-        }
-    }
-}
-
-dependencies {
-    implementation(libs.androidx.activity.compose)
-
-    implementation(projects.core.designsystem)
-    implementation(projects.core.ui)
-}
-
-dependencyGuard {
-    configuration("releaseRuntimeClasspath")
-}
