--- app-geto-catalog/src/main/res/drawable/ic_launcher_foreground.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-  ~
-  ~   Copyright 2023 Einstein Blanco
-  ~
-  ~   Licensed under the GNU General Public License v3.0 (the "License");
-  ~   you may not use this file except in compliance with the License.
-  ~   You may obtain a copy of the License at
-  ~
-  ~       https://www.gnu.org/licenses/gpl-3.0
-  ~
-  ~   Unless required by applicable law or agreed to in writing, software
-  ~   distributed under the License is distributed on an "AS IS" BASIS,
-  ~   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~   See the License for the specific language governing permissions and
-  ~   limitations under the License.
-  ~
-  -->
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:width="108dp"
-    android:height="108dp"
-    android:viewportWidth="108"
-    android:viewportHeight="108">
-    <path
-        android:fillColor="#fefefe"
-        android:pathData="M65.86,54.57c-0.02,0.53 -0.07,1.03 -0.26,1.5 -0.41,1.01 -1.15,1.57 -2.25,1.66 -0.43,0.03 -0.85,0 -1.26,-0.12 -0.79,-0.23 -1.3,-0.77 -1.6,-1.51 -0.25,-0.62 -0.3,-1.28 -0.26,-1.94 0.03,-0.49 0.13,-0.96 0.35,-1.4 0.4,-0.79 1.03,-1.26 1.91,-1.38 0.43,-0.06 0.86,-0.05 1.28,0.04 0.79,0.17 1.34,0.63 1.7,1.34 0.24,0.46 0.34,0.96 0.37,1.48 0.01,0.11 0.01,0.22 0.01,0.34zM73.97,60.47c0.06,1.49 0.13,2.97 0.19,4.44 0.04,1 0.08,2 0.12,3.01 0.05,1.18 0.11,2.37 0.16,3.55 0.01,0.16 0.01,0.32 0.02,0.48 0.04,0.86 0.07,1.72 0.1,2.58a6.97,6.97 0,0 1,0.01 0.42c-0.03,1.37 -0.66,2.4 -1.83,3.09 -0.49,0.29 -1.04,0.43 -1.62,0.44 -0.46,0.01 -0.91,-0.06 -1.36,-0.17 -0.21,-0.05 -0.42,-0.09 -0.63,-0.13 -0.44,-0.09 -0.88,-0.18 -1.32,-0.27 -0.4,-0.08 -0.81,-0.15 -1.21,-0.23 -0.38,-0.08 -0.75,-0.16 -1.12,-0.24 -0.84,-0.16 -1.67,-0.37 -2.53,-0.33a6.77,6.77 0,0 0,-1.18 0.16c-0.36,0.08 -0.72,0.15 -1.08,0.23 -0.35,0.08 -0.71,0.17 -1.06,0.25 -0.35,0.08 -0.71,0.15 -1.06,0.23 -0.37,0.08 -0.75,0.17 -1.12,0.25 -0.34,0.08 -0.68,0.15 -1.02,0.22 -0.26,0.06 -0.51,0.11 -0.77,0.16 -0.11,0.02 -0.23,0.04 -0.34,0.04 -0.88,0 -1.75,0.01 -2.63,-0 -0.27,-0.01 -0.54,-0.07 -0.81,-0.13 -0.4,-0.08 -0.8,-0.17 -1.2,-0.26 -0.35,-0.08 -0.69,-0.16 -1.04,-0.24 -0.19,-0.04 -0.38,-0.09 -0.57,-0.13 -0.54,-0.12 -1.09,-0.24 -1.63,-0.36a77.06,77.06 0,0 0,-1.08 -0.23c-0.2,-0.04 -0.41,-0.08 -0.61,-0.11a5.2,5.2 0,0 0,-1.79 -0c-0.5,0.09 -0.99,0.19 -1.48,0.28 -0.37,0.07 -0.74,0.14 -1.11,0.22 -0.41,0.08 -0.83,0.17 -1.24,0.26 -0.24,0.05 -0.48,0.09 -0.71,0.14 -0.23,0.05 -0.46,0.1 -0.69,0.14 -0.48,0.09 -0.95,0.19 -1.43,0.25a3.06,3.06 0,0 1,-1.24 -0.1c-0.84,-0.25 -1.51,-0.72 -2,-1.44 -0.41,-0.61 -0.62,-1.29 -0.61,-2.03 0.01,-0.4 0.02,-0.79 0.04,-1.19 0.02,-0.53 0.05,-1.06 0.07,-1.59 0.05,-1.18 0.11,-2.37 0.16,-3.55 0.03,-0.73 0.06,-1.47 0.09,-2.2 0.04,-1.09 0.08,-2.18 0.13,-3.27 0.05,-1.12 0.1,-2.23 0.15,-3.35 0.05,-1.06 0.09,-2.13 0.13,-3.19 0.05,-1.22 0.11,-2.45 0.16,-3.67 0.05,-1.1 0.09,-2.19 0.13,-3.29 0.03,-0.79 0.06,-1.58 0.12,-2.36 0.05,-0.61 0.15,-1.22 0.24,-1.82 0.06,-0.4 0.13,-0.81 0.23,-1.21 0.13,-0.5 0.27,-1 0.43,-1.49a15.21,15.21 0,0 1,0.41 -1.14c0.19,-0.47 0.39,-0.94 0.61,-1.4 0.49,-1.01 1.07,-1.97 1.73,-2.88 0.85,-1.17 1.82,-2.23 2.91,-3.18 0.93,-0.82 1.94,-1.54 3.01,-2.16 0.83,-0.49 1.69,-0.91 2.59,-1.26a23.78,23.78 0,0 1,1.69 -0.59c0.44,-0.13 0.9,-0.23 1.35,-0.33 0.41,-0.09 0.83,-0.18 1.25,-0.25 0.27,-0.05 0.55,-0.08 0.82,-0.1 1.17,-0.1 2.35,-0.17 3.53,-0.05 0.47,0.05 0.94,0.08 1.41,0.14 0.23,0.03 0.45,0.07 0.68,0.11 0.43,0.09 0.86,0.16 1.28,0.28 0.49,0.13 0.98,0.27 1.46,0.44 0.52,0.18 1.03,0.38 1.54,0.59 0.66,0.27 1.29,0.6 1.91,0.95 1.41,0.81 2.7,1.78 3.87,2.91a19.19,19.19 0,0 1,2.79 3.41,19.31 19.31,0 0,1 1.53,2.9c0.28,0.66 0.52,1.33 0.73,2 0.13,0.42 0.23,0.84 0.32,1.27 0.09,0.41 0.17,0.81 0.25,1.22 0.04,0.23 0.08,0.46 0.1,0.7 0.05,0.53 0.1,1.06 0.13,1.59 0.09,1.91 0.18,3.82 0.26,5.73 0.04,1.01 0.08,2.02 0.12,3.03 0.05,1.26 0.12,2.53 0.17,3.81m-3.67,-7.45c-0.07,-0.29 -0.13,-0.58 -0.22,-0.86 -0.42,-1.32 -1.16,-2.42 -2.22,-3.31 -1.17,-0.97 -2.52,-1.53 -4.04,-1.67 -0.97,-0.09 -1.92,-0.02 -2.86,0.24 -1.18,0.33 -2.23,0.91 -3.12,1.76 -0.88,0.84 -1.51,1.84 -1.88,3 -0.43,1.37 -0.48,2.75 -0.16,4.15 0.43,1.88 1.47,3.37 3.08,4.42 1.72,1.12 3.61,1.44 5.61,1.07 1.73,-0.32 3.19,-1.17 4.31,-2.54 1.03,-1.26 1.54,-2.71 1.62,-4.33 0.03,-0.63 -0.01,-1.26 -0.13,-1.92m-18.56,0.8h-6.82c-0.05,0.23 -0.04,2.89 0.01,3.05h2.24c0.19,0 0.38,-0 0.57,0 0.18,0 0.37,0 0.59,0.01 -0.47,0.6 -1.02,0.99 -1.72,1.13 -0.68,0.14 -1.36,0.09 -2.01,-0.12 -0.73,-0.23 -1.22,-0.71 -1.5,-1.42 -0.14,-0.36 -0.21,-0.74 -0.24,-1.12 -0.05,-0.66 -0.02,-1.31 0.23,-1.93 0.25,-0.64 0.67,-1.12 1.32,-1.37 0.62,-0.23 1.26,-0.24 1.9,-0.06 0.35,0.1 0.65,0.29 0.87,0.59 0.08,0.11 0.16,0.14 0.29,0.14 1.38,-0 2.76,-0 4.14,-0h0.2c0,-0.07 0.01,-0.12 -0,-0.16 -0.16,-0.95 -0.52,-1.82 -1.15,-2.57 -0.8,-0.95 -1.83,-1.53 -3.02,-1.84 -0.87,-0.22 -1.75,-0.29 -2.64,-0.25 -0.95,0.04 -1.87,0.22 -2.74,0.62 -1.48,0.66 -2.53,1.75 -3.17,3.24 -0.33,0.77 -0.51,1.58 -0.57,2.42 -0.1,1.33 0.06,2.62 0.59,3.85 0.39,0.92 0.97,1.71 1.75,2.34 0.89,0.72 1.9,1.16 3.02,1.36 1.01,0.18 2.02,0.18 3.04,0.02 0.7,-0.11 1.38,-0.31 2,-0.66 1.2,-0.68 2.16,-1.61 2.9,-2.78 0.07,-0.11 0.1,-0.22 0.1,-0.36 -0,-1.3 -0,-2.6 -0,-3.9 0,-0.05 -0,-0.09 -0,-0.14 -0,-0.08 -0.06,-0.08 -0.15,-0.08m0.02,17.41c-0,-0.96 -0.01,-1.92 -0.01,-2.89 0,-0.01 -0.01,-0.01 -0.01,-0.02 -0,-0.01 -0.01,-0.01 -0.01,-0.01 -0.43,-0.04 -2.17,-0.02 -2.31,0.02 -0.04,0.39 -0.02,6.1 0.02,6.25h9.38v-6.27h-2.34v2.21c0,0.19 0,0.38 -0,0.57 -0.01,0.18 0.03,0.36 -0.02,0.54h-1.31v-2.9L52.92,68.72v2.9h-1.15v-0.39m5.55,-27.29h0.94c0.01,-0.04 0.03,-0.05 0.03,-0.07a2.44,2.44 0,0 0,0.01 -0.18c-0,-1.16 -0.01,-2.33 -0.01,-3.49 0,-0.01 -0.01,-0.01 -0.01,-0.02a0.22,0.22 0,0 0,-0.01 -0.01c-0.46,-0.04 -8.47,-0.02 -8.61,0.02v3.72c0.02,0.01 0.04,0.02 0.05,0.02 2.53,0 5.06,0.01 7.63,0.01m6.18,0h4.37c0.02,-0.03 0.03,-0.04 0.03,-0.05a0.39,0.39 0,0 0,0.01 -0.08c0,-1.18 0,-2.36 -0,-3.53 0,-0.03 -0.01,-0.05 -0.01,-0.08 -0,-0.01 -0.01,-0.01 -0.01,-0.02a0.25,0.25 0,0 0,-0.01 -0.01c-0.45,-0.04 -8.47,-0.02 -8.61,0.02v3.75h4.25M39.95,40.34v3.59h8.65v-3.77h-8.37c-0.05,0 -0.11,0 -0.16,0 -0.11,-0.01 -0.12,0.06 -0.12,0.18m16.69,20.66c-0.02,-0.11 -0.11,-0.07 -0.18,-0.07a1875.85,1875.85 0,0 0,-4.82 -0c-0.05,0 -0.1,0.01 -0.16,0.01v1.52h1.6v4.59h1.92v-3.28c0,-0.22 -0,-0.44 0,-0.67 0,-0.21 -0.01,-0.43 0.01,-0.65h1.63c0,-0.48 0,-0.94 -0,-1.44z"
-        tools:ignore="VectorPath" />
-</vector>
