+++ core/screenshot-testing/src/main/kotlin/com/android/geto/core/screenshot/testing/util/ScreenshotHelper.kt
@@ -18,30 +18,15 @@
 package com.android.geto.core.screenshot.testing.util
 
 import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.key
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.platform.LocalInspectionMode
-import androidx.compose.ui.test.ExperimentalTestApi
-import androidx.compose.ui.test.hasTestTag
 import androidx.compose.ui.test.junit4.AndroidComposeTestRule
 import androidx.compose.ui.test.onRoot
 import androidx.test.ext.junit.rules.ActivityScenarioRule
-import com.android.geto.core.designsystem.theme.GetoTheme
 import com.github.takahirom.roborazzi.ExperimentalRoborazziApi
 import com.github.takahirom.roborazzi.RoborazziOptions
 import com.github.takahirom.roborazzi.RoborazziOptions.CompareOptions
 import com.github.takahirom.roborazzi.RoborazziOptions.RecordOptions
 import com.github.takahirom.roborazzi.captureRoboImage
-import com.github.takahirom.roborazzi.captureScreenRoboImage
-import com.google.accompanist.testharness.TestHarness
-import org.robolectric.RuntimeEnvironment
 
 @OptIn(ExperimentalRoborazziApi::class)
 val DefaultRoborazziOptions = RoborazziOptions(
@@ -52,8 +37,7 @@ val DefaultRoborazziOptions = RoborazziOptions(
 )
 
 enum class DefaultTestDevices(val description: String, val spec: String) {
-    PHONE("phone", "spec:shape=Normal,width=640,height=360,unit=dp,dpi=480"),
-    FOLDABLE(
+    PHONE("phone", "spec:shape=Normal,width=640,height=360,unit=dp,dpi=480"), FOLDABLE(
         "foldable",
         "spec:shape=Normal,width=673,height=841,unit=dp,dpi=480",
     ),
@@ -74,126 +58,25 @@ fun <A : ComponentActivity> AndroidComposeTestRule<ActivityScenarioRule<A>, A>.c
     }
 }
 
-fun <A : ComponentActivity> AndroidComposeTestRule<ActivityScenarioRule<A>, A>.captureMultiDeviceSnackbar(
-    snackbarHostState: SnackbarHostState,
-    message: String,
-    testTag: String,
-    fileName: String,
-    body: @Composable () -> Unit,
-) {
-    DefaultTestDevices.entries.forEach {
-        this.captureSnackbarForDevice(
-            snackbarHostState = snackbarHostState,
-            message = message,
-            testTag = testTag,
-            deviceName = it.description,
-            deviceSpec = it.spec,
-            fileName = fileName,
-            body = body,
-        )
-    }
-}
-
 fun <A : ComponentActivity> AndroidComposeTestRule<ActivityScenarioRule<A>, A>.captureForDevice(
-    deviceName: String,
-    deviceSpec: String,
-    fileName: String,
-    roborazziOptions: RoborazziOptions = DefaultRoborazziOptions,
-    darkMode: Boolean = false,
-    body: @Composable () -> Unit,
-) {
-    val (width, height, dpi) = extractSpecs(deviceSpec)
-
-    // Set qualifiers from specs
-    RuntimeEnvironment.setQualifiers("w${width}dp-h${height}dp-${dpi}dpi")
-
-    this.activity.setContent {
-        CompositionLocalProvider(
-            LocalInspectionMode provides true,
-        ) {
-            TestHarness(darkMode = darkMode) {
-                body()
-            }
-        }
-    }
-
-    this.onRoot().captureRoboImage(
-        filePath = "src/test/screenshots/${fileName}_$deviceName.png",
-        roborazziOptions = roborazziOptions,
-    )
-}
-
-@OptIn(ExperimentalTestApi::class)
-fun <A : ComponentActivity> AndroidComposeTestRule<ActivityScenarioRule<A>, A>.captureSnackbarForDevice(
-    snackbarHostState: SnackbarHostState,
-    message: String,
-    testTag: String,
-    deviceName: String,
-    deviceSpec: String,
-    fileName: String,
-    roborazziOptions: RoborazziOptions = DefaultRoborazziOptions,
-    darkMode: Boolean = false,
-    body: @Composable () -> Unit,
-) {
-    val (width, height, dpi) = extractSpecs(deviceSpec)
-
-    // Set qualifiers from specs
-    RuntimeEnvironment.setQualifiers("w${width}dp-h${height}dp-${dpi}dpi")
-
-    this.activity.setContent {
-        CompositionLocalProvider(
-            LocalInspectionMode provides true,
-        ) {
-            TestHarness(darkMode = darkMode) {
-                LaunchedEffect(key1 = true) {
-                    snackbarHostState.showSnackbar(message = message)
-                }
-
-                body()
-            }
-        }
-    }
-
-    waitUntilAtLeastOneExists(matcher = hasTestTag(testTag = testTag))
-
-    this.onRoot().captureRoboImage(
-        filePath = "src/test/screenshots/${fileName}_$deviceName.png",
-        roborazziOptions = roborazziOptions,
-    )
-}
-
-/**
- * Experimental feature to capture the entire screen, including dialogs combining light/dark and default/Android themes and whether dynamic color
- * is enabled.
- */
-@OptIn(ExperimentalRoborazziApi::class)
-fun <A : ComponentActivity> AndroidComposeTestRule<ActivityScenarioRule<A>, A>.captureDialogForDevice(
-    name: String,
     fileName: String,
     deviceName: String,
     deviceSpec: String,
-    roborazziOptions: RoborazziOptions = DefaultRoborazziOptions,
     darkMode: Boolean = false,
     body: @Composable () -> Unit,
 ) {
-    val (width, height, dpi) = extractSpecs(deviceSpec)
-
-    // Set qualifiers from specs
-    RuntimeEnvironment.setQualifiers("w${width}dp-h${height}dp-${dpi}dpi")
-
-    this.activity.setContent {
-        CompositionLocalProvider(
-            LocalInspectionMode provides true,
-        ) {
-            TestHarness(darkMode = darkMode) {
-                body()
-            }
-        }
-    }
-
-    captureScreenRoboImage(
-        filePath = "src/test/screenshots/$name/${fileName}_$deviceName.png",
-        roborazziOptions = roborazziOptions,
+    captureForDeviceUtil(
+        fileName = fileName,
+        deviceName = deviceName,
+        deviceSpec = deviceSpec,
+        darkMode = darkMode,
+        body = body,
+        capture = { filePath, roborazziOptions ->
+            this.onRoot().captureRoboImage(
+                filePath = filePath,
+                roborazziOptions = roborazziOptions,
+            )
+        },
     )
 }
 
@@ -209,68 +92,23 @@ fun <A : ComponentActivity> AndroidComposeTestRule<ActivityScenarioRule<A>, A>.c
     shouldCompareAndroidTheme: Boolean = true,
     content: @Composable (desc: String) -> Unit,
 ) {
-    val darkModeValues = if (shouldCompareDarkMode) listOf(true, false) else listOf(false)
-    val dynamicThemingValues = if (shouldCompareDynamicColor) listOf(true, false) else listOf(false)
-    val androidThemeValues = if (shouldCompareAndroidTheme) listOf(true, false) else listOf(false)
-
-    var darkMode by mutableStateOf(true)
-    var dynamicTheming by mutableStateOf(false)
-    var androidTheme by mutableStateOf(false)
-
-    this.setContent {
-        CompositionLocalProvider(
-            LocalInspectionMode provides true,
-        ) {
-            GetoTheme(
-                androidTheme = androidTheme,
-                darkTheme = darkMode,
-                disableDynamicTheming = !dynamicTheming,
-            ) {
-                // Keying is necessary in some cases (e.g. animations)
-                key(androidTheme, darkMode, dynamicTheming) {
-                    val description = generateDescription(
-                        shouldCompareDarkMode = shouldCompareDarkMode,
-                        darkMode = darkMode,
-                        shouldCompareAndroidTheme = shouldCompareAndroidTheme,
-                        androidTheme = androidTheme,
-                        shouldCompareDynamicColor = shouldCompareDynamicColor,
-                        dynamicTheming = dynamicTheming,
-                    )
-                    content(description)
-                }
-            }
-        }
-    }
-
-    // Create permutations
-    darkModeValues.forEach { isDarkMode ->
-        darkMode = isDarkMode
-        val darkModeDesc = if (isDarkMode) "dark" else "light"
-
-        androidThemeValues.forEach { isAndroidTheme ->
-            androidTheme = isAndroidTheme
-            val androidThemeDesc = if (isAndroidTheme) "androidTheme" else "defaultTheme"
-
-            dynamicThemingValues.forEach dynamicTheme@{ isDynamicTheming ->
-                // Skip tests with both Android Theme and Dynamic color as they're incompatible.
-                if (isAndroidTheme && isDynamicTheming) return@dynamicTheme
-
-                dynamicTheming = isDynamicTheming
-                val dynamicThemingDesc = if (isDynamicTheming) "dynamic" else "notDynamic"
-
-                val filename = overrideFileName ?: name
-
-                this.onRoot().captureRoboImage(
-                    filePath = "src/test/screenshots/" + "$name/$filename" + "_$darkModeDesc" + "_$androidThemeDesc" + "_$dynamicThemingDesc" + ".png",
-                    roborazziOptions = DefaultRoborazziOptions,
-                )
-            }
-        }
-    }
+    captureMultiThemeUtil(
+        name = name,
+        overrideFileName = overrideFileName,
+        shouldCompareDarkMode = shouldCompareDarkMode,
+        shouldCompareDynamicColor = shouldCompareDynamicColor,
+        shouldCompareAndroidTheme = shouldCompareAndroidTheme,
+        content = content,
+        capture = { filePath, roborazziOptions ->
+            this.onRoot().captureRoboImage(
+                filePath = filePath, roborazziOptions = roborazziOptions,
+            )
+        },
+    )
 }
 
 @Composable
-private fun generateDescription(
+internal fun generateDescription(
     shouldCompareDarkMode: Boolean,
     darkMode: Boolean,
     shouldCompareAndroidTheme: Boolean,
@@ -298,7 +136,7 @@ private fun generateDescription(
 /**
  * Extracts some properties from the spec string. Note that this function is not exhaustive.
  */
-private fun extractSpecs(deviceSpec: String): TestDeviceSpecs {
+internal fun extractSpecs(deviceSpec: String): TestDeviceSpecs {
     val specs = deviceSpec.substringAfter("spec:").split(",").map { it.split("=") }
         .associate { it[0] to it[1] }
     val width = specs["width"]?.toInt() ?: 640
