+++ settings.gradle.kts
@@ -35,7 +35,6 @@ dependencyResolutionManagement {
 enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 rootProject.name = "Geto"
 include(":app")
-include(":app-geto-catalog")
 include(":benchmarks")
 include(":core:clipboardmanager")
 include(":core:common")
